fix(server): NextFreeNodeID use int loop var; assert MAC in node-id-set test
- node_registry.go: loop variable changed from uint8 to int to prevent
  wrap-around infinite loop when all 255 IDs are in use
- server_enrollment_test.go: assert Data[1:7] matches enrolled node MAC
  in TestApproveEnrollment_SendsNodeIdSet_AfterJoinAck

+++ server/orchestrator/mesh/node_registry.go
@@ -92,9 +92,9 @@ func (nr *NodeRegistry) NextFreeNodeID() uint8 {
 			used[n.NodeID] = true
 		}
 	}
-	for id := uint8(1); id <= 255; id++ {
-		if !used[id] {
-			return id
+	for id := 1; id <= 255; id++ {
+		if !used[uint8(id)] {
+			return uint8(id)
 		}
 	}
 	return 0 // all 255 IDs in use
